requirements.txt
asgiref==3.5.2
bcrypt==4.0.0
DateTime==4.7
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
zope.interface==5.4.0
